requirements.txt
# Add your Python package dependencies here
# Example:
# requests>=2.28.0
# pandas>=2.0.0
# numpy>=1.24.0
langchain-huggingface==1.2.0
mcp-use==1.5.2
python-dotenv==1.2.1
truststore
aiohttp